requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
psycopg2-binary==2.9.9
django-filter==23.5
django-cors-headers==4.3.1
python-dotenv==1.0.1
openai==1.12.0
drf-yasg==1.21.7
cryptography==41.0.7
celery==5.3.6
redis==5.0.1
django-storages==1.14.2
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0 